requirements.txt
Flask
Flask-Cors
python-dotenv
pandas
openpyxl
PyJWT
google-generativeai
faiss-cpu    # if your environment allows; else comment it out and keyword fallback works
numpy
langgraph
duckduckgo_search
sentence-transformers
scikit-learn
xgboost
dash
dash_bootstrap_components
requests
ddgs
langchain_community
serpapi
google-search-results
flask
flask-cors
python-dotenv
pandas
numpy
langchain
langgraph
langchain-google-genai
google-generativeai
duckduckgo_search
langchain_huggingface